etl/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
redis==5.0.1
numpy==1.24.3